Restructure DataLoader following comments
