+++ configs/config.yaml
@@ -6,3 +6,4 @@
 # model:
 #   hidden_units: 128
 #   learning_rate: 0.001
+batch_size: 64
