+++ requirements.txt
@@ -5,3 +5,4 @@ torch
 pandas
 torchvision
 tensorboard
+pyyaml
