+++ src/utils.py
@@ -3,6 +3,18 @@ import zipfile
 import tarfile
 import shutil
 from pathlib import Path
+import yaml
+
+def list_files_in_folder(folder_path):
+    try:
+        # Get a list of all files and directories in the folder
+        files_in_folder = os.listdir(folder_path)
+        # Show only jpg files
+        files = [file for file in files_in_folder if file.lower().endswith('.jpg')]
+        return files
+    except Exception as e:
+        print("Error listing files in folder:", e)
+        return None
 
 def UnZipFolder(file_path, folder_inside, destination_path):
     # Unzipp a folder from inside a ZIP file to a destination
@@ -51,3 +63,21 @@ def UnZipFolder(file_path, folder_inside, destination_path):
 
             except Exception as e:
                 print("Error extracting from TAR/TGZ file:", e)
+
+def read_yaml(file_path):
+    """
+    Reads a YAML file and returns its contents as a dictionary.
+    
+    :param file_path: Path to the YAML file.
+    :return: Contents of the YAML file as a dictionary.
+    """
+    try:
+        with open(file_path, 'r') as file:
+            data = yaml.safe_load(file)
+            return data
+    except FileNotFoundError:
+        print(f"Error: The file {file_path} was not found.")
+    except yaml.YAMLError as exc:
+        print(f"Error parsing YAML file: {exc}")
+    except Exception as exc:
+        print(f"An unexpected error occurred: {exc}")
